requirements.txt
alembic==1.16.4
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
behave==1.2.6
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
cssselect==1.3.0
dash==3.0.4
dash-bootstrap-components==2.0.2
dash-extensions==2.0.4
dash-table==5.0.0
dataclass-wizard==0.35.0
EditorConfig==0.17.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.18.0
Flask==3.0.3
Flask-Caching==2.3.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-ads==26.1.0
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleads==44.0.0
googleapis-common-protos==1.66.0
greenlet==3.2.4
grpcio==1.72.0rc1
grpcio-status==1.71.0
gspread==6.1.4
h11==0.16.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
# Editable Git install with no remote (line_item_details==0.0.0)
-e c:\users\amit.parmar\desktop\display_qc
lxml==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
more-itertools==10.7.0
mycdp==1.2.0
mypy_extensions==1.1.0
narwhals==1.38.2
nest-asyncio==1.6.0
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pathspec==0.12.1
pdbp==1.7.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pynose==1.5.4
pyotp==2.9.0
pyparsing==3.2.1
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rsa==4.9
sbvirtualdisplay==1.4.0
selenium==4.32.0
seleniumbase==4.38.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
tabcompleter==1.4.0
tabulate==0.9.0
tk==0.1.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.45.1
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.14.2
zeep==4.3.1
zipp==3.21.0
